Define a plugin for semantic search indexing at build time
Create a plugin that will iterate over all the documents and use the
Universal Sentence Encoder (USE) model to generate embeddings for each
document chunk.

This plugin should correspond to a search bar component that uses USE to
generate an embedding of a user query and uses some similarity metric
like kNN or cosine similarity to identify the matching document chunks
and resolve the matching documents.